--split test into AttributesConfigTests and AttributesManagersTests
This split separates godly test file into tests primarily around configuration functionality, such as setting values properly and saving and loading consistency, and manager functionality, such as create, delete, copy, etc.
